* deactivate/activate and context menu functionality for backup screen

+++ src/main/java/tpiskorski/vboxcm/ui/controller/backup/BackupController.java
@@ -1,21 +1,25 @@
 package tpiskorski.vboxcm.ui.controller.backup;
 
-import tpiskorski.vboxcm.config.Config;
-import tpiskorski.vboxcm.config.ConfigService;
-import tpiskorski.vboxcm.core.backup.BackupDefinition;
-import tpiskorski.vboxcm.core.backup.BackupDefinitionService;
-import tpiskorski.vboxcm.ui.core.ContextAwareSceneLoader;
 import javafx.beans.binding.Bindings;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.geometry.Pos;
 import javafx.scene.control.Button;
+import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextField;
 import javafx.scene.control.*;
+import javafx.scene.input.MouseButton;
+import javafx.scene.input.MouseEvent;
 import javafx.stage.Stage;
 import javafx.util.Callback;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import tpiskorski.vboxcm.config.Config;
+import tpiskorski.vboxcm.config.ConfigService;
+import tpiskorski.vboxcm.core.backup.BackupDefinition;
+import tpiskorski.vboxcm.core.backup.BackupDefinitionService;
+import tpiskorski.vboxcm.ui.control.ConfirmationAlertFactory;
+import tpiskorski.vboxcm.ui.core.ContextAwareSceneLoader;
 
 import java.awt.*;
 import java.io.File;
@@ -40,8 +44,26 @@ public class BackupController {
 
     @Autowired private ModifyVmBackupController modifyVmBackupController;
 
+    @FXML private ContextMenu contextMenu;
+    @FXML private MenuItem dynamicMenuItem;
+
     @FXML
     public void initialize() throws IOException {
+        backupsTableView.addEventHandler(MouseEvent.MOUSE_CLICKED, t -> {
+            if (t.getButton() == MouseButton.SECONDARY) {
+                BackupDefinition selectedItem = backupsTableView.getSelectionModel().getSelectedItem();
+                if (selectedItem.isActive()) {
+                    dynamicMenuItem.setText("Deactivate");
+                    dynamicMenuItem.setOnAction(this::deactivate);
+                } else {
+                    dynamicMenuItem.setText("Activate");
+                    dynamicMenuItem.setOnAction(this::activate);
+                }
+
+                contextMenu.show(backupsTableView, t.getScreenX(), t.getScreenY());
+            }
+        });
+
         backupLocation.setText(configService.getConfig().getBackupLocation());
         configService.addPropertyChangeListener(evt -> {
             Config newValue = (Config) evt.getNewValue();
@@ -91,6 +113,30 @@ public class BackupController {
         backupsTableView.setItems(backupDefinitionService.getBackups());
     }
 
+    private void deactivate(ActionEvent actionEvent) {
+        boolean confirmed = ConfirmationAlertFactory.createAndShow(
+            "Do you really want to deactivate this backup?",
+            "Watchdog"
+        );
+
+        if (confirmed) {
+            BackupDefinition backupToDeactivate = backupsTableView.getSelectionModel().getSelectedItem();
+            backupDefinitionService.deactivate(backupToDeactivate);
+        }
+    }
+
+    private void activate(ActionEvent actionEvent) {
+        boolean confirmed = ConfirmationAlertFactory.createAndShow(
+            "Do you really want to activate this backup?",
+            "Watchdog"
+        );
+
+        if (confirmed) {
+            BackupDefinition backupToDeactivate = backupsTableView.getSelectionModel().getSelectedItem();
+            backupDefinitionService.activate(backupToDeactivate);
+        }
+    }
+
     private void browseHomeDirectory() {
         try {
             Desktop.getDesktop().open(new File(System.getProperty("user.home")));
@@ -109,12 +155,19 @@ public class BackupController {
 
     @FXML
     public void removeVm() {
-        BackupDefinition backupDefinitionToRemove = backupsTableView.getSelectionModel().getSelectedItem();
-        backupDefinitionService.remove(backupDefinitionToRemove);
+        boolean confirmed = ConfirmationAlertFactory.createAndShow(
+            "Do you really want to remove this backup?",
+            "Watchdog"
+        );
+
+        if (confirmed) {
+            BackupDefinition backupDefinitionToRemove = backupsTableView.getSelectionModel().getSelectedItem();
+            backupDefinitionService.remove(backupDefinitionToRemove);
+        }
     }
 
     @FXML
-    public void modify() {
+    public void modifyVm() {
         if (modifyVmStage.isShowing()) {
             modifyVmBackupController.clear();
             modifyVmStage.hide();
